refactor: Rating 컴포넌트 모노레포에 최신화
