+++ monorepo/packages/ui/src/constants/rating.ts
@@ -1,0 +1,1 @@
+export const TOTAL_STARS_DEFAULT = 5;
